Reorganize senders
